Update bearing calculation
Caused by the changed coordinate system from NUE to ENU, the example code to calculate the bearing angle also has to be updated. In the commit 120524ebce053b395c77d174335c38684e96682e the description was updated to ENU, but the code wasnt. This missing update was found during a bug discussion on discord (https://discord.com/channels/565154702715518986/565154703139405824/966431158638379039)